corrected cd leaks and invalid read (of size 1) reproduced some weird comportements of __

+++ lib/libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ahenault <ahenault@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/05 14:48:51 by maemaldo          #+#    #+#             */
-/*   Updated: 2024/10/14 19:18:15 by ahenault         ###   ########.fr       */
+/*   Updated: 2024/10/15 20:44:40 by ahenault         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s && s[i])
